requirements.txt
flask==2.3.3
langchain==0.1.7
langchain-google-genai==0.0.9
python-dotenv==1.0.0
flask-session==0.5.0 